pyproject.toml
[project]
name = "me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.3.28",
    "langchain-text-splitters>=0.3.9",
    "openai>=1.97.1",
    "pinecone>=5,<6",
    "pydantic>=2.11.7",
    "pypdf>=5.9.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "tiktoken>=0.9.0",
    "uvicorn>=0.35.0",
]
